HTBHF-2663 Move VerificationResult into the model package as it will be returned from the controller.
